split into printful-lib and printful-cli
- extract `printful_lib::Printful` and `get_store`
- minimise dependencies
