+++ Cargo.toml
@@ -1,15 +1,6 @@
-[package]
-name = "printful_api"
-version = "0.1.0"
-authors = ["Mike Moran <mike@houseofmoran.com>"]
-edition = "2018"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-reqwest = { version = "0.10", features = ["json"] }
-tokio = { version = "0.2", features = ["full"] }
-dotenv = "0.15"
-serde_json = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-base64 = "0.12"
+members = [
+    "cli",
+    "lib",
+]
