+++ cli/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "printful-cli"
+version = "0.1.0"
+authors = ["Mike Moran <mike@houseofmoran.com>"]
+edition = "2018"
+
+[dependencies]
+reqwest = { version = "0.10", features = ["json"] }
+tokio = { version = "0.2", features = ["full"] }
+dotenv = "0.15"
+
+printful-lib = { path = "../lib"}
